# Conflicts:
#	Four Peaks/AlphaTemp/main.cpp
#	Four Peaks/AlphaTemp/tutorial.cpp
#	Four Peaks/Assets/Levels/autumn_s3.txt

+++ Four Peaks/AlphaTemp/splash.cpp
@@ -57,7 +57,7 @@ namespace game
     // -------------------------------------------------------------------------
     int SplashScreen::update()
     {
-        float dt = AEFrameRateControllerGetFrameTime();
+        float dt = static_cast<float>(AEFrameRateControllerGetFrameTime());
         timer += dt;
 
         // Allow the player to skip the splash early
